Services, Models, Repositories added and WeatherForecast logic moved to new Service
